ETSWORK-9967 - Add Sign In and No Network screens used for authentication
The `SignIn` component is used to allow customers to provide, change, and reset company code. It also provides a sign in button for the user to kick off the authentication process. For transparency due to browser constraints and the way authentication works we have to have a sign in button that a user clicks.

The `NetworkSplash` in a component that renders it's children if the network is available. It's meant to only be used in the authentication process per requirements. We will likely implement a different model for showing offline status when in the application in the future.

Other changes of note:

* Had to change TypeScript output to `ES2017` to align with Capacitor v3 targets
* Changed TypeScript configuration to output type definitions so it can be more easily consumed in the app
* Needed to change the build process so image assets are part of the module so they can be used in the app
* Added `@testing-library/jest-dom` to allow more simplified jest matchers
* Added ESLint plugins for `jest-dom`, `@testing-library/react` and `eslint-plugin-react` to enforce best practices