ETSWORK-10017 - Adds a screen for choosing / switching spaces
This screen will primarily be used in the initial user setup where they will need to choose a space to work in. This screen can also be used when a user is switching spaces as well. The Radio's needed to be updated to support our use case here primarily being able to provide captions on both the radio label and item label. Button needed to be adjusted to support expanding to the full-width of its container.

I also had to make an adjustment to `preview.js` in this PR. In my last PR I made some tweaks to address font rendering issues between iOS and Android. As part of that I removed the `mode` setting. Unfortunately this caused other rendering to break between platforms so I needed to add it back.

Finally I cleaned up some of the lint warnings.