ETSWORK-3069 - Adds the Smart Notification Detail screen
The Smart Notifications Detail screen includes a number of different pieces. Arguably the most complex component is the dynamic rendering of content as requested by the connector developer. We make use of `rehype-react` for the rendering based on the "Workgrid AST" that will be delivered as a HAST compatible tree via the API (a change that still needs to be implemented). With `rehype-react` we provide a React component for every "element" in our AST. All Smart Notification components in the project are prefixed with `SN` to clearly identify their intended use. The various `SNInput*` and button components wrap the respective public input and button components with the necessary constraints and defaults for adaptive card as discussed with @team/ux. Form validation has been implemented with `react-hook-form` and a custom resolver that constructs a `yup` schema. A custom resolver was required because not all actions on Smart Notifications require validation.

This work is also the first introduction of our "Screen Pattern" to support both small and large screens within the constraints of Ionic. You can find more details and patterns in the `README`.

All screens have been renamed to use the new postfix `Screen` naming convention so it's clear what public components are screens and which are not.