ETSWORK-10471 - Convert `@testing-library/react` and `JSDOM` tests to use Cypress component tester.
There were a number of reasons for the conversion to Cypress component tester but the two major reasons were:

1. Ionic is based on Web Components and `JSDOM` has very limited support
2. Both `@testing-library/react` amd Ionic must fake events which reduces confidence that components will work in practice.

I decided to go with a `.cypress.tsx` syntax so we can isolate Cypress tests from the other tests such as image snapshots and A11y tests. This will also future proof us if we ever add plain old unit tests that can still run with Jest.

The build has been updated accordingly with an additional step to ensure Cypress tests pass with screenshots and videos being uploaded as artifacts for failures.